refactor(crdt): DRY — one home for the wire codec + doc-key constants
DRY pass over the new Relay-model modules vs the code they replace. Two
primitives were duplicated:
  - frame codec (MESSAGE_SYNC/toB64/fromB64/encodeUpdateFrame) — was in both
    channel.ts and note-provider.ts -> extracted to src/crdt/wire.ts; both
    import it (channel re-exports the public names for its importers).
  - Y.Doc key constants + frontmatterOf/rawFrontmatterOf — was in both
    manager.ts and provider-registry.ts -> moved to frontmatter-codec.ts (the
    pure codec home); manager re-exports for back-compat; registry imports the
    shared accessors and drops its inline projection.

No behavior change. Full suite 2235 green.
